backend/requirements.txt
python-dotenv 
google-generativeai
fastapi
uvicorn